// src/types/next-auth.d.ts
import type { DefaultSession } from 'next-auth';

export type UserRole = 'user' | 'investor' | 'admin';

declare module 'next-auth' {
  interface Session {
    user: {
      id: string;
      role: UserRole;
    } & DefaultSession['user'];
  }

  interface User {
    role?: UserRole;
  }
}

declare module 'next-auth/jwt' {
  interface JWT {
    id?: string;
    role?: UserRole;
  }
}